Finished Project
Apologies for the 1 commit I had problems with github desktop and I didnt want to risk losing my project but ill hapilly take the deducted marks for the 1 commit again my sincerest apologies.

This project is a parcel management website with ads attatched to it the admins can delete parcel information and change ads and users can come in to check the contents of their parcel.